refactor(spec): split chumsky_parser.rs into per-concern submodules
Move-only split of the 4,596-line `chumsky_parser.rs` into a directory
module with per-concern submodules (lex, expr, decls, handler,
instruction, interface, tests) behind a `mod.rs` facade that keeps the
top-level driver (`spec_parser` / `parse` / `format_parse_error`).

The directory rename preserves the module path
`crate::spec::chumsky_parser` and the root re-export
`crate::chumsky_parser`; internal cross-references resolve via
`pub(in crate::spec::chumsky_parser)` re-export globs. No behavior or
generated-output changes — all snapshot suites and the 1006-test unit
suite pass unchanged.

Mirrors the chumsky_adapter split (#112).
